feat: integrate dummy api to edit profile screen
